fix: resolve JSX runtime in telegram-guide test (closes #397)
